pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "p2"
version = "0.1.0"
description = "S3-compatible object storage system"
requires-python = ">=3.12"
dependencies = [
    # Django core
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-filter>=23.0",
    "django-crispy-forms>=2.0",
    "django-extensions>=3.2",
    "django-debug-toolbar>=4.0",

    # Database
    "psycopg[binary]>=3.1",

    # Auth
    "djangorestframework-simplejwt>=5.3",
    "authlib>=1.3",

    # API schema
    "drf-spectacular>=0.27",

    # Image processing
    "Pillow>=10.0",

    # ASGI server
    "uvicorn[standard]>=0.27",
    "uvloop>=0.19",

    # Async task queue
    "arq>=0.25",

    # Redis
    "redis[hiredis]>=5.0",

    # Async I/O
    "aiofiles>=23.0",
    "aiobotocore>=2.9",

    # gRPC
    "grpcio>=1.60",
    "grpcio-reflection>=1.60",
    "protobuf>=5.0",

    # OpenTelemetry
    "opentelemetry-sdk>=1.22",
    "opentelemetry-exporter-otlp>=1.22",
    "opentelemetry-instrumentation-django>=0.43b0",
    "opentelemetry-instrumentation-logging>=0.43b0",

    # Encryption
    "cryptography>=42.0",

    # Utilities
    "boto3>=1.34",
    "PyYAML>=6.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-django>=4.8",
    "hypothesis>=6.100",
    "fakeredis[aioredis]>=2.21",
    "moto[s3]>=5.0",

    # Coverage
    "coverage>=7.0",

    # Linting / formatting
    "isort>=5.13",
    "pylint>=3.0",
    "pylint-django>=2.5",
    "prospector>=1.10",
    "autopep8>=2.0",
    "colorama>=0.4",

    # Dev tools
    "django-extensions>=3.2",
    "bumpversion>=0.6",
    "grpcio-tools>=1.60",
    "unittest-xml-reporting>=3.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-django>=4.8",
    "hypothesis>=6.100",
    "fakeredis[aioredis]>=2.21",
    "moto[s3]>=5.0",
    "coverage>=7.0",
    "isort>=5.13",
    "pylint>=3.0",
    "pylint-django>=2.5",
    "prospector>=1.10",
    "autopep8>=2.0",
    "colorama>=0.4",
    "bumpversion>=0.6",
    "grpcio-tools>=1.60",
    "unittest-xml-reporting>=3.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
DJANGO_SETTINGS_MODULE = "p2.core.settings"

[tool.isort]
profile = "django"
known_django = ["django"]
known_third_party = ["rest_framework", "drf_spectacular", "arq", "redis", "opentelemetry"]

[tool.coverage.run]
source = ["p2"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
